weapon_selector_requirements.txt
# Sujuon Weapon Selector Dependencies
PyGithub>=1.59
docker>=7.0
apscheduler>=3.10
requests>=2.31
